browser-runtime/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic-settings>=2.2
httpx==0.28.1
pytest==8.3.5
playwright>=1.49.0
